Establish rough content and features
Updated the README.md to include a small introduction on the premise of
the project and the knowledge I will utilise.

Added extremely barebones layout to the website without much design.
Then added in content to populate the page to see the design better and
wrote functions in script.js that creates new book cards and code that
opens the dialog box.

Next up, I will configure the dialog box and the form to facilitate the
adding of new books to the display and then add the necessary buttons as
per project requirements, before finally working on a rough design for
the website.