Add dialog and remove functionalities
Add the dialog algorithm to allow for user to provide their own input
and be able to see it in the display afterwards, as well as the ability
to remove the display.

Adjusted multiple code to make it neater and more efficient, such as
shifting code blocks and removing unnecessary or repeated code.

Add CSS for some elements to get a slightly more aesthetic website.