Add read button and creation of display on loading
Changed html and js to allow the cards to be formed upon loading through
the library array instead of merely existing in the html.

Read button added which will toggle between read and not read state, css
has yet to be implemented in yet.

The overall functionality is all there, leaving only the design to be
worked on.